+++ src/util/validator-schema.js
@@ -1,4 +1,4 @@
-import validations from "@/config/imports/validation-rules_test";
+import validations from "@/config/imports/schema-rules";
 import { v4 as uuid } from "uuid";
 
 const Serializer = new XMLSerializer();
@@ -20,10 +20,10 @@ export default {
         const scopedTests = validations.filter(v => v.assert).filter(v => !scope || scope === v.scope);
         const checkedSchemaRules = scopedTests.length;
         for (let test of scopedTests) {
-            const selection = document.querySelectorAll(test.path);
-            const pass = test.assert(Array.from(selection), document);
+            const selection = Array.from(document.querySelectorAll(test.path));
+            const pass = test.assert(selection, document);
             if (!pass) {
-                const result = (test.getInvalid) ? test.getInvalid(Array.from(selection), document) : [];
+                const result = (test.getInvalid) ? test.getInvalid(selection, document) : [];
                 if (result.length) {
                     for (let element of result) {
                         const { location, lineNr, lines } = this.getLocation(element, lineMap, lineArr);
@@ -66,13 +66,7 @@ export default {
     }
 };
 
-export function iirdsValidateXmlRule(rdfDoc, iirdsRule) {
-    const selection = Array.from(rdfDoc.querySelectorAll(iirdsRule.path));
-    const findInvalidElements = iirdsRule?.findInvalidElements ?? (() => []);
-    const assert = iirdsRule?.assert ?? ((selElements, invalidElements) => ((invalidElements?.length ?? 0) === 0));
-    const invalidElements = findInvalidElements(selection);
-    return { succeeded: assert(selection, invalidElements), invalidElements };
+export function validateSingleRule(document, rule) {
+    const selection = Array.from(document.querySelectorAll(test.path));
+    return { succeeded: rule?.assert(selection, document) || true, invalidElements: rule?.getInvalid(selection) };
 }
-
-export const IIRDS_RULES = validations;
-
